Move Firebase Cloud Function to the impl package
The ...impl package is ignored by the test coverage check
