+++ app/src/main/java/ch/epfl/sweng/eventmanager/repository/impl/FirebaseCloudFunction.java
@@ -1,0 +1,45 @@
+package ch.epfl.sweng.eventmanager.repository.impl;
+
+import com.google.android.gms.tasks.Continuation;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.functions.FirebaseFunctions;
+import com.google.firebase.functions.HttpsCallableResult;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import androidx.annotation.NonNull;
+
+public class FirebaseCloudFunction {
+    /**
+     * Calls a dedicated FireBase Cloud Function allowing an event administrator to add an user to
+     * its event.
+     *
+     * @param email email of the target user
+     * @param eventId target event
+     * @param role string representation role to be assigned to the target user
+     * @return the related task
+     */
+    public static Task<Boolean> addUserToEvent(String email, int eventId, String role) {
+        // Prepare parameters for the Firebase Cloud Function
+        Map<String, Object> data = new HashMap<>();
+        data.put("eventId", eventId);
+        data.put("userEmail", email);
+        data.put("role", role);
+        data.put("push", true);
+
+        return FirebaseFunctions.getInstance()
+                .getHttpsCallable("addUserToEvent")
+                .call(data)
+                .continueWith(new Continuation<HttpsCallableResult, Boolean>() {
+                    @Override
+                    public Boolean then(@NonNull Task<HttpsCallableResult> task) throws Exception {
+                        // This continuation runs on either success or failure, but if the task
+                        // has failed then getResult() will throw an Exception which will be
+                        // propagated down.
+                        Boolean result = (Boolean) task.getResult().getData();
+                        return result;
+                    }
+                });
+    }
+}
